Remove the .file-drop-hover class when the drag leaves the dropzone
Fixes a bug where the class stays applied if the drag ends outside the
dropzone.  This isn't symptomatic in the Methylation Station usage
because the whole page is the dropzone, so it was impossible for the
drag to end outside of it.

Switching to dragleave means we also need to check if the element being
left is our element or not, since our dragleave handler will trigger on
fully-contained child elements of our element.