Fix another edge case around the removal of the .file-drop-hover class
The commit "Remove the .file-drop-hover class when the drag leaves the
dropzone" (c8f24b6) was good, but didn't quite go far enough.  When the
drag enters a child element of the dropzone, a dragenter is fired for
the child (which is fine) but a dragleave is _also_ fired for the
dropzone (i.e. the containing element).  We need to ignore that case
which is identifiable by checking if the element being entered when we
"leave" the dropzone is indeed a child of the dropzone.

I'd previously accomplished this (and forgotten about doing so) in our
upload-list component by using CSS to prevent mouse events
(pointer-events: none) on an active dropzone, but this is a more elegant
solution which doesn't require supplementary CSS.